requirements.txt
appdirs==1.4.3
asgiref==3.2.5
aspy.yaml==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.10b0
cfgv==3.1.0
click==7.1.1
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.11
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
nodeenv==1.3.5
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
pre-commit==1.21.0
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pytest==4.6.9
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
virtualenv==20.0.10
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.1.0
